// fe/src/lib/api/generated/model/civilStatusCategoryValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CivilStatusCategoryValue = typeof CivilStatusCategoryValue[keyof typeof CivilStatusCategoryValue];


export const CivilStatusCategoryValue = {
  SINGLE: 'SINGLE',
  MARRIED: 'MARRIED',
  SOLO_PARENT: 'SOLO_PARENT',
  OTHER: 'OTHER',
  NOT_SPECIFIED: 'NOT_SPECIFIED',
} as const;