// fe/src/lib/api/generated/model/courseChoiceReasonValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CourseChoiceReasonValue = typeof CourseChoiceReasonValue[keyof typeof CourseChoiceReasonValue];


export const CourseChoiceReasonValue = {
  INTEREST_APTITUDE: 'INTEREST_APTITUDE',
  MINIMAL_COST: 'MINIMAL_COST',
  FRIENDS: 'FRIENDS',
  PARENT_CHOICE: 'PARENT_CHOICE',
  JOB_OPPORTUNITIES: 'JOB_OPPORTUNITIES',
  OTHER: 'OTHER',
} as const;