// fe/src/lib/api/generated/model/educationLevelValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type EducationLevelValue = typeof EducationLevelValue[keyof typeof EducationLevelValue];


export const EducationLevelValue = {
  PREPARATORY: 'PREPARATORY',
  ELEMENTARY: 'ELEMENTARY',
  JUNIOR_HIGH: 'JUNIOR_HIGH',
  SENIOR_HIGH: 'SENIOR_HIGH',
  TECHNICAL_VOCATIONAL: 'TECHNICAL_VOCATIONAL',
  COLLEGIATE: 'COLLEGIATE',
} as const;