// fe/src/lib/api/generated/model/fourPsStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type FourPsStatusValue = typeof FourPsStatusValue[keyof typeof FourPsStatusValue];


export const FourPsStatusValue = {
  BENEFICIARY: 'BENEFICIARY',
  NOT_BENEFICIARY: 'NOT_BENEFICIARY',
  NOT_SPECIFIED: 'NOT_SPECIFIED',
} as const;